+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/common/jwt/JwtUtil.java
@@ -1,6 +1,6 @@
 package com.sparta.blackwhitedeliverydriver.common.jwt;
 
-import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.UserRoleEnum;
 import com.sparta.blackwhitedeliverydriver.common.exception.CustomJwtException;
 import com.sparta.blackwhitedeliverydriver.common.exception.JwtExceptionMessage;
 import io.jsonwebtoken.Claims;
