+++ blackWhiteDeliveryDriver/src/test/java/com/sparta/blackwhitedeliverydriver/bascket/service/BasketServiceTest.java
@@ -9,20 +9,20 @@ import static org.mockito.Mockito.when;
 
 import com.sparta.blackwhitedeliverydriver.bascket.dto.BasketAddRequestDto;
 import com.sparta.blackwhitedeliverydriver.bascket.dto.BasketResponseDto;
-import com.sparta.blackwhitedeliverydriver.bascket.entity.Basket;
-import com.sparta.blackwhitedeliverydriver.order.entity.Order;
-import com.sparta.blackwhitedeliverydriver.product.entity.Product;
-import com.sparta.blackwhitedeliverydriver.store.entity.Store;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
-import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.model.bascket.entity.Basket;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.Order;
+import com.sparta.blackwhitedeliverydriver.model.product.entity.Product;
+import com.sparta.blackwhitedeliverydriver.model.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.UserRoleEnum;
 import com.sparta.blackwhitedeliverydriver.bascket.exception.BasketExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.order.exception.OrderExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.bascket.repository.BasketRepository;
-import com.sparta.blackwhitedeliverydriver.order.repository.OrderRepository;
-import com.sparta.blackwhitedeliverydriver.product.repository.ProductRepository;
-import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.model.bascket.repository.BasketRepository;
+import com.sparta.blackwhitedeliverydriver.model.order.repository.OrderRepository;
+import com.sparta.blackwhitedeliverydriver.model.product.repository.ProductRepository;
+import com.sparta.blackwhitedeliverydriver.model.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.model.user.repository.UserRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
