+++ blackWhiteDeliveryDriver/src/test/java/com/sparta/blackwhitedeliverydriver/order/controller/OrderControllerTest.java
@@ -20,9 +20,9 @@ import com.sparta.blackwhitedeliverydriver.order.dto.OrderGetDetailResponseDto;
 import com.sparta.blackwhitedeliverydriver.order.dto.OrderGetResponseDto;
 import com.sparta.blackwhitedeliverydriver.order.dto.OrderResponseDto;
 import com.sparta.blackwhitedeliverydriver.order.dto.OrderUpdateRequestDto;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderStatusEnum;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderTypeEnum;
-import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderStatusEnum;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderTypeEnum;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.UserRoleEnum;
 import com.sparta.blackwhitedeliverydriver.mock.user.MockUser;
 import com.sparta.blackwhitedeliverydriver.order.service.OrderService;
 import java.util.List;
