+++ blackWhiteDeliveryDriver/src/test/java/com/sparta/blackwhitedeliverydriver/order/service/OrderServiceTest.java
@@ -12,23 +12,23 @@ import com.sparta.blackwhitedeliverydriver.order.dto.OrderAddRequestDto;
 import com.sparta.blackwhitedeliverydriver.order.dto.OrderGetDetailResponseDto;
 import com.sparta.blackwhitedeliverydriver.order.dto.OrderResponseDto;
 import com.sparta.blackwhitedeliverydriver.order.dto.OrderUpdateRequestDto;
-import com.sparta.blackwhitedeliverydriver.bascket.entity.Basket;
-import com.sparta.blackwhitedeliverydriver.order.entity.Order;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderProduct;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderStatusEnum;
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderTypeEnum;
-import com.sparta.blackwhitedeliverydriver.product.entity.Product;
+import com.sparta.blackwhitedeliverydriver.model.bascket.entity.Basket;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.Order;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderProduct;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderStatusEnum;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderTypeEnum;
+import com.sparta.blackwhitedeliverydriver.model.product.entity.Product;
 import com.sparta.blackwhitedeliverydriver.pay.service.PayService;
-import com.sparta.blackwhitedeliverydriver.store.entity.Store;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
-import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.model.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.UserRoleEnum;
 import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.order.exception.OrderExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.bascket.repository.BasketRepository;
-import com.sparta.blackwhitedeliverydriver.order.repository.OrderProductRepository;
-import com.sparta.blackwhitedeliverydriver.order.repository.OrderRepository;
-import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.model.bascket.repository.BasketRepository;
+import com.sparta.blackwhitedeliverydriver.model.order.repository.OrderProductRepository;
+import com.sparta.blackwhitedeliverydriver.model.order.repository.OrderRepository;
+import com.sparta.blackwhitedeliverydriver.model.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.model.user.repository.UserRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
